main dashboard has a button that takes you to where you can connect to plaid
