+++ src/components/SignIn/SignIn.js
@@ -5,6 +5,7 @@ import { Alert } from "@material-ui/lab";
 import TextField from "@material-ui/core/TextField";
 import Link from "@material-ui/core/Link";
 import Grid from "@material-ui/core/Grid";
+import Button from "@material-ui/core/Button";
 import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
@@ -46,7 +47,8 @@ function SignIn(props) {
     });
   };
 
-  const onHandleLogin = async () => {
+  const onHandleLogin = async (e) => {
+    e.preventDefault();
     try {
       const res = await axios.post(
         "http://localhost:3301/users/login",
@@ -109,9 +111,9 @@ function SignIn(props) {
               Your email or password was incorrect.
             </Alert>
           ) : null}
-          <div className={classes.submit} onClick={onHandleLogin}>
+          <Button className={classes.submit} onClick={onHandleLogin}>
             Sign In
-          </div>
+          </Button>
           <Grid container>
             <Grid item xs>
               <Link href="#" variant="body2">
